[products] moving lifx specific definition code out of registry
